criar função para atualizar o status do emprestimo

+++ controllers/emprestimo.py
@@ -8,32 +8,62 @@ from flask_login import login_required, current_user
 
 emprestimo_bp = Blueprint('emprestimo', __name__, template_folder='../templates/emprestimos')
 
-@emprestimo_bp.route('/visualizar_emprestimos', methods=['GET'])
+def definir_status(data_prazo: date, data_devolucao: date | None) -> str:
+    data_atual = date.today()
+
+    if data_devolucao:
+        return "Devolvido"
+
+    if data_atual > data_prazo:
+        return "Atrasado"
+
+    return "Pendente"
+
+
+@emprestimo_bp.route('/listar_emprestimos', methods=['GET'])
 @login_required
-def visualizar_emprestimos():
+def listar_emprestimos():
     with Session(bind=engine) as db:
         emprestimos = db.query(Emprestimos).where(Emprestimos.user_id == current_user.id).all()
-        return render_template('emprestimos/visualizar.html', emprestimos = emprestimos)
+
+        # atualizar status
+        for emprestimo in emprestimos:
+            novo_status = definir_status(emprestimo.data_prazo, emprestimo.data_devolucao)
+            if novo_status != emprestimo.status:
+                emprestimo.status = novo_status
+
+        db.commit()
+
+        return render_template('emprestimos/listar.html', emprestimos = emprestimos)
 
 
 @emprestimo_bp.route('/register_emprestimo/<int:livro_id>')
 @login_required
 def register_emprestimo(livro_id: int):
-    user_id = current_user.id
-    data_atual = date.today()
-    data_prazo = data_atual + timedelta(days=30)
     with Session(bind=engine) as db:
-        novo_emprestimo = Emprestimos(data_emprestimo=data_atual, data_prazo=data_prazo, user_id=user_id, livro_id=livro_id)
-        db.add(novo_emprestimo)
-        db.commit()
-    
-    with Session(bind=engine) as db:
-        livro = db.get(Livros, livro_id)
-        livro.quantidade -= 1
-        db.commit()
 
-    flash('Empréstimo realizado com sucesso!', category='success')
-    return redirect(url_for('emprestimo.visualizar_emprestimos'))
+        verificar_livro = db.query(Livros).filter_by(id = livro_id).first()
+    
+    if verificar_livro and verificar_livro.quantidade > 0:
+        user_id = current_user.id
+        data_atual = date.today()
+        data_prazo = data_atual + timedelta(days=30)
+        with Session(bind=engine) as db:
+            novo_emprestimo = Emprestimos(data_emprestimo=data_atual, data_prazo=data_prazo, user_id=user_id, livro_id=livro_id)
+            db.add(novo_emprestimo)
+            db.commit()
+        
+        with Session(bind=engine) as db:
+            livro = db.get(Livros, livro_id)
+            livro.quantidade -= 1
+            db.commit()
+        
+        flash('Empréstimo realizado com sucesso!', category='success')
+        return redirect(url_for('emprestimo.listar_emprestimos'))
+    
+    flash('O empréstimo não pode ser realizado!', category='error')
+    return redirect(url_for('livro.listar_livros'))
+
     
 
 @emprestimo_bp.route('/editar_emprestimo/<int:emprestimo_id>', methods=['GET', 'POST'])
@@ -50,7 +80,7 @@ def editar_emprestimo(emprestimo_id: int):
         with Session(bind=engine) as db:
             if emprestimo.status == 'Devolvido':
                 flash("A ação não pode ser realizada!", category='error')
-                return redirect(url_for('emprestimo.visualizar_emprestimos'))
+                return redirect(url_for('emprestimo.listar_emprestimos'))
             emprestimo = db.get(Emprestimos, emprestimo_id)
             emprestimo.data_emprestimo = data_emprestimo
             emprestimo.data_prazo = data_prazo
@@ -59,7 +89,7 @@ def editar_emprestimo(emprestimo_id: int):
             db.commit()
         
         flash('Empréstimo editado com sucesso!', category='success')
-        return redirect(url_for('emprestimo.visualizar_emprestimos'))
+        return redirect(url_for('emprestimo.listar_emprestimos'))
     
     return render_template('emprestimos/editar.html', emprestimo=emprestimo)
 
@@ -89,18 +119,18 @@ def devolver_livro(emprestimo_id):
         db.commit()
 
     flash('O livro foi devolvido', category='success')
-    return redirect(url_for("emprestimo.visualizar_emprestimos"))
+    return redirect(url_for("emprestimo.listar_emprestimos"))
 
 
-@emprestimo_bp.route('/rm_emprestimo/<int:emprestimo_id>', methods=['GET', 'POST'])
+@emprestimo_bp.route('/deletar_emprestimo/<int:emprestimo_id>', methods=['GET', 'POST'])
 @login_required
-def rm_emprestimo(emprestimo_id):
+def deletar_emprestimo(emprestimo_id):
     with Session(bind=engine) as db:
         emprestimo = db.get(Emprestimos, emprestimo_id)
-        if emprestimo.status == 'Devolvido':
+        if emprestimo and emprestimo.status == 'Devolvido':
             db.delete(emprestimo)
             db.commit()
             flash("Emprestimo deletado com sucesso", category='succes')
         flash("A ação não pode ser realizada!", category='error')
-    return redirect(url_for('emprestimo.visualizar_emprestimos'))
+    return redirect(url_for('emprestimo.listar_emprestimos'))
